backend/server/requirements.txt
flask
flask_Cors
python-dotenv
qdrant_client
sentence_transformers
